Test 'inlining' case
This is related to the discussion of https://github.com/trailofbits/siderophile/pull/17

We had to remove `-C inline-threshold=9999` from `RUSTFLAGS` we
pass to `cargo rustc ...` when emitting llvm. Otherwise, the compiler
inlined a function call and we didn't get proper results from
Siderophile. The test added in here checks if we find proper badness of
this particular example.