Add workaround for crates with rust-toolchain file
There is an option to specify `rust-toolchain` file for your crate,
to tell cargo which toolchain version to use.

Normal builds via `cargo build` compiles the crate and its dependencies with
one toolchain version and so does not respect dependencies
`rust-toolchain` files.

However, Siderophile somehow compiles packages one by one and do indeed
respect the `rust-toolchain` file, leading to compiling dependencies
with wrong toolchain versions (if they are installed) or not compiling
them at all. In the end this ends up failing the compilation (no matter
if the depedency was compiled with wrong toolchain version) and so
Siderophile fails to do its job.

This commit adds a workaround for this issue by simply specifying the
`RUSTUP_TOOLCHAIN` environment variable that overrides the toolchain
version with the default one present in the system.

See more information on https://github.com/trailofbits/siderophile/issues/14